Clean.

+++ seika/src/rendering/renderer.h
@@ -15,7 +15,7 @@ extern "C" {
 
 void se_renderer_initialize(int inWindowWidth, int inWindowHeight, int inResolutionWidth, int inResolutionHeight);
 void se_renderer_finalize();
-void se_renderer_update_window_size(float windowWidth, float windowHeight);
+void se_renderer_update_window_size(int windowWidth, int windowHeight);
 void se_renderer_set_sprite_shader_default_params(SEShader* shader);
 void se_renderer_queue_sprite_draw_call(SETexture* texture, SERect2 sourceRect, SESize2D destSize, SEColor color, bool flipH, bool flipV, SETransformModel2D* globalTransform, int zIndex, SEShaderInstance* shaderInstance);
 void se_renderer_queue_font_draw_call(SEFont* font, const char* text, float x, float y, float scale, SEColor color, int zIndex);
